+++ com/authorize/http/request/webmessage.js
@@ -2,4 +2,4 @@ exports = module.exports = function() {
   return require('oauth2orize-wmrm').extensions();
 };
 
-exports['@implements'] = 'http://i.authnomicon.org/oauth2/authorization/http/RequestParameters';
+exports['@implements'] = 'module:oauth2orize.RequestParametersProcessor';